- Some refactoring - Started working on result data component
